% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defenestrate-package.R
\docType{package}
\name{defenestrate-package}
\alias{defenestrate}
\alias{defenestrate-package}
\title{defenestrate: Helpers for Working with Microsoft's Online Services Programmatically}
\description{
Some functions to simplify the process of working with files on Microsoft's online teamworking products (OneDrive, Teams). Built largely around the 'Microsoft365R' package.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
