% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_auth}
\alias{.check_auth}
\title{Check for a Specified Token}
\usage{
.check_auth(scope_group = c("teams", "business_onedrive"))
}
\arguments{
\item{scope_group}{Character. The name for a set of scopes that allow access
to Microsoft services. See details.}
}
\value{
Nothing if the specified token is found, otherwise an error.
}
\description{
Check for a Specified Token
}
\details{
Each Microsoft product requires a token that covers particular
scopes of interaction. The names of these, as used in the \code{scope_group}
argument, are analogous to the associated functions in 'Microsoft365R':
\code{"team"} for using Teams, the token for which is generated when you use
\link[Microsoft365R:client]{Microsoft365R::list_teams}; \code{"business_onedrive"} for using OneDrive, the
token for which is generated when you use
\link[Microsoft365R:client]{Microsoft365R::get_business_onedrive}.
}
