% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.extract_file_type}
\alias{.extract_file_type}
\title{Identify File Type and Check that It's Valid}
\usage{
.extract_file_type(filepath)
}
\arguments{
\item{filepath}{Character. The path to a file that contains tabular data to
be read in.}
}
\value{
A character string representing the filetype.
}
\description{
Identify File Type and Check that It's Valid
}
