% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.read_file_type}
\alias{.read_file_type}
\title{Choose a Function to Read a Given File Type and Read It}
\usage{
.read_file_type(file_type, temp_file, ...)
}
\arguments{
\item{file_type}{Character. The file extension of the file being read.}

\item{temp_file}{Character. A filepath to a temporary file where the data
will be written.}

\item{...}{Additional arguments to pass to the function that will be used to
read in the data, which depends on the type of file being read.}
}
\value{
A data.frame/tibble if the file exists and can be read.
}
\description{
Choose a Function to Read a Given File Type and Read It
}
