% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{list_onedrive_files}
\alias{list_onedrive_files}
\title{List Available Files in a Business OneDrive Directory}
\usage{
list_onedrive_files(filepath = NULL)
}
\arguments{
\item{filepath}{Character. The path to a OneDrive directory starting from the
root of your business OneDrive directory. Defaults to \code{NULL}, which will
return the directories and files available to you from the root.}
}
\value{
Character. The names of directories (suffixed with a forward slash)
and files (suffixed with a file extension) available from the provided
filepath.
}
\description{
A shortcut for listing files in a directory in a business OneDrive that you
have access to and have been authorised against (see details) without needing
to interact with the R6 object returned by
\link[Microsoft365R:client]{Microsoft365R::get_business_onedrive}.
}
\details{
You must have completed the authorisation process for the channel
that you want to read from. To do this, run
\code{Microsoft365R::get_business_onedrive()}, which will open the browser for
automatic authorisation. A token authorising you for the required scopes
will be stored on your computer.

Note that you'll see the message 'Loading Microsoft Graph login for
default tenant' when tokens are being checked. You may also see 'Access
token has expired or is no longer valid; refreshing' when your token is
stale. Use \link[base:message]{base::suppressMessages} to hide these messages.
}
\examples{
\dontrun{list_onedrive_files("Personal")}
}
