% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{list_teams_channels}
\alias{list_teams_channels}
\title{List Available Channels in a Teams Team}
\usage{
list_teams_channels(team)
}
\arguments{
\item{team}{Character. The name of a Teams team.}
}
\value{
Character. Names of Teams channels available to you.
}
\description{
A shortcut for listing channels in a Teams team that you have access to and
have been authorised against (see details) without needing to interact with
the R6 object returned by \link[Microsoft365R:client]{Microsoft365R::get_team}.
}
\details{
You must have completed the authorisation process for Teams. To do
this, run \code{Microsoft365R::list_teams()}, which will open the browser for
automatic authorisation. A token authorising you for the required scopes
will be stored on your computer.

Note that you'll see the message 'Loading Microsoft Graph login for
default tenant' when tokens are being checked. You may also see 'Access
token has expired or is no longer valid; refreshing' when your token is
stale. Use \link[base:message]{base::suppressMessages} to hide these messages.
}
\examples{
\dontrun{list_teams_channels("Statistics Production Division"))}
}
