% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_channel_table}
\alias{read_channel_table}
\title{Read Tabular Data from a Teams Channel}
\usage{
read_channel_table(
  team = "Statistics Production Division",
  channel = "Team files and A_L",
  filepath = "QPD/Knowledge bank/QPD knowledge bank.xlsx",
  ...
)
}
\arguments{
\item{team}{Character. The name of a Teams team.}

\item{channel}{Character. The name of a channel in the Teams team provided to
the 'team' argument.}

\item{filepath}{Character. The path to a file that contains tabular data to
be read in.}

\item{...}{Additional arguments to pass to the function that will be used to
read in the data, which depends on the type of file being read. See
details.}
}
\value{
A data.frame/tibble class object.
}
\description{
Read in data from a file that's available from a channel in a Microsoft Teams
team that you have access to and have been authorised against (see details).
}
\details{
You need to have a token on your computer that authorises you to use
the scopes required to access data on Teams. The easiest way to do this is
to run a function like \link[Microsoft365R:client]{Microsoft365R::list_teams}, which will open a
browser window and authorise you automatically. Then you can use
{defenestrate} functions. You only need to do this authorisation once.

Note that you'll see the message 'Loading Microsoft Graph login for
default tenant' when tokens are being checked. You may also see 'Access
token has expired or is no longer valid; refreshing' when your token is
stale. Use \link[base:message]{base::suppressMessages} to hide these messages.

The exact function used internally by \link{read_channel_table} to read
data depends on the filetype. For Excel files, \link[readxl:read_excel]{readxl::read_excel}; for
ODS files, \link[readODS:read_ods]{readODS::read_ods}; for CSV and TSV files, \link[readr:read_delim]{readr::read_csv}
and \link[readr:read_delim]{readr::read_tsv}; for RDS files, \link[readr:read_rds]{readr::read_rds}. You can supply
extra arguments from these functions to the '...' (dots) argument. For
example, if reading an xlsx file, you could use the argument \code{sheet = 1} to
read the first sheet only.
}
\examples{
    \dontrun{read_channel_table("Team Name", "Channel Name", "file.csv")}
}
